chore: factor out convertToScriptUpdate logic
